begin function logic
* parser: new function declaration parsing
* add symbol and improve memory table
* add __str__ methode for printing a memory table
* interpreter: function declaration typing and determination of the result
* interpreter: function call typing

+++ lexer.py
@@ -99,7 +99,7 @@ class Lexer:
             result += self.current_char
             self.advance()
         
-        return Token(FLOAT, int(result))
+        return Token(FLOAT, float(result))
     
     def get_string_token(self):
         """
